+++ actix-http/CHANGES.md
@@ -1,6 +1,17 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+### Added
+* `AnyBody::empty` for quickly creating an empty body. [#2446]
+
+### Changed
+* Rename `AnyBody::{Message => Stream}`. [#2446]
+
+### Removed
+* `AnyBody::Empty`; an empty body can now only be represented as a zero-length `Bytes` variant. [#2446]
+* `BodySize::Empty`; an empty body can now only be represented as a `Sized(0)` variant. [#2446]
+
+[#2446]: https://github.com/actix/actix-web/pull/2446
 
 
 ## 3.0.0-beta.12 - 2021-11-15